+++ magik-squid/src/test/java/nl/ramsolutions/sw/magik/analysis/typing/reasoner/recursive/ParameterRecursiveReasonerTest.java
@@ -1,0 +1,3 @@
+package nl.ramsolutions.sw.magik.analysis.typing.reasoner.recursive;
+
+class ParameterRecursiveReasonerTest {}
